+++ libc/src/spawn/posix_spawn_file_actions_adddup2.cpp
@@ -10,6 +10,7 @@
 
 #include "file_actions.h"
 #include "src/__support/CPP/new.h"
+#include "src/__support/alloc-checker.h"
 #include "src/__support/common.h"
 #include "src/__support/libc_errno.h"
 #include "src/__support/macros/config.h"
